Move function to portin module
